Remove history integration
- Expose only the ScrollBehavior class
- Remove explicit history dependencies
