+++ karma.conf.js
@@ -1,10 +1,10 @@
 const webpack = require('webpack'); // eslint-disable-line import/no-extraneous-dependencies
 
-module.exports = config => {
+module.exports = (config) => {
   const { env } = process;
 
   config.set({
-    frameworks: ['mocha'],
+    frameworks: ['mocha', 'sinon-chai'],
 
     files: ['test/index.js'],
