+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "scroll-behavior",
   "version": "0.8.2",
-  "description": "Scroll management for history",
+  "description": "Pluggable browser scroll management",
   "files": [
     "es",
     "lib"
@@ -24,8 +24,6 @@
     "url": "git+https://github.com/taion/scroll-behavior.git"
   },
   "keywords": [
-    "history",
-    "location",
     "scroll"
   ],
   "author": "Jimmy Jia",
@@ -35,40 +33,41 @@
   },
   "homepage": "https://github.com/taion/scroll-behavior#readme",
   "dependencies": {
-    "dom-helpers": "^2.4.0",
+    "dom-helpers": "^3.0.0",
     "invariant": "^2.2.1"
   },
-  "peerDependencies": {
-    "history": "^1.12.1 || ^2.0.0"
-  },
   "devDependencies": {
-    "babel-cli": "^6.11.4",
-    "babel-core": "^6.13.2",
-    "babel-eslint": "^6.1.2",
-    "babel-loader": "^6.2.4",
+    "babel-cli": "^6.18.0",
+    "babel-core": "^6.18.2",
+    "babel-eslint": "^7.1.0",
+    "babel-loader": "^6.2.7",
     "babel-plugin-add-module-exports": "^0.2.1",
     "babel-plugin-dev-expression": "^0.2.1",
-    "babel-plugin-istanbul": "^1.0.3",
-    "babel-polyfill": "^6.13.0",
-    "babel-preset-es2015": "^6.13.2",
-    "babel-preset-stage-1": "^6.13.0",
+    "babel-plugin-istanbul": "^2.0.3",
+    "babel-polyfill": "^6.16.0",
+    "babel-preset-latest": "^6.16.0",
+    "babel-preset-stage-2": "^6.18.0",
     "chai": "^3.5.0",
     "codecov": "^1.0.1",
-    "cross-env": "^2.0.0",
-    "eslint": "^3.2.2",
-    "eslint-config-airbnb-base": "^5.0.1",
-    "eslint-plugin-import": "^1.12.0",
+    "cross-env": "^3.1.3",
+    "dirty-chai": "^1.2.2",
+    "eslint": "^3.9.1",
+    "eslint-config-4catalyzer": "^0.1.3",
+    "eslint-plugin-import": "^1.16.0",
     "history": "^2.1.2",
-    "karma": "^1.1.2",
-    "karma-chrome-launcher": "^1.0.1",
+    "karma": "^1.3.0",
+    "karma-chrome-launcher": "^2.0.0",
     "karma-coverage": "^1.1.1",
     "karma-firefox-launcher": "^1.0.0",
-    "karma-mocha": "^1.1.1",
-    "karma-mocha-reporter": "^2.1.0",
+    "karma-mocha": "^1.2.0",
+    "karma-mocha-reporter": "^2.2.0",
+    "karma-sinon-chai": "^1.2.4",
     "karma-sourcemap-loader": "^0.3.7",
-    "karma-webpack": "^1.7.0",
-    "mocha": "^3.0.1",
+    "karma-webpack": "^1.8.0",
+    "mocha": "^3.1.2",
     "rimraf": "^2.5.4",
-    "webpack": "^1.13.1"
+    "sinon": "^1.17.6",
+    "sinon-chai": "^2.8.0",
+    "webpack": "^1.13.3"
   }
 }
