+++ test/.eslintrc
@@ -2,9 +2,11 @@
   "env": {
     "mocha": true
   },
+  "globals": {
+    "expect": false
+  },
   "rules": {
-    "no-unused-expressions": 0,
-    "import/no-extraneous-dependencies": [2, {
+    "import/no-extraneous-dependencies": ["error", {
       "devDependencies": true
     }]
   }
