+++ test/index.js
@@ -1,5 +1,9 @@
 import 'babel-polyfill';
 
+import dirtyChai from 'dirty-chai';
+
+global.chai.use(dirtyChai);
+
 // Ensure all files in src folder are loaded for proper code coverage analysis.
 const srcContext = require.context('../src', true, /.*\.js$/);
 srcContext.keys().forEach(srcContext);
